Application/requirements.txt
altgraph==0.17.5
packaging==25.0
psycopg==3.2.13
psycopg-binary==3.2.13
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dotenv==1.2.1
setuptools==80.9.0
shiboken6==6.10.1
